+++ piezas/bishop.py
@@ -1,28 +1,16 @@
 from piezas.piece import Piece
 
 class Bishop(Piece):
+   white_str = "♗"
+   black_str = "♝"
    def __init__(self, color):
         super().__init__(color)
+        
+    
 
-   def pieza_del_mismo_color(self, row, col, board):
-        #Esta funcion sirve para ver si hay otra pieza del mismo color
-        pieza = board.get_piece(row,col)
-        if pieza is not None and pieza.__color__ == self.__color__:
-            return True
-        else:
-            return False
-        
-   def captura(self, row, col, board):
-        #Verifica si la torre puede capturar 
-        pieza = board.get_piece(row,col)
-        if pieza is not None and pieza.__color__ != self.__color__:
-            return True
-        else:
-            return False
-        
+    #Veridica los movimientos diagonales
    def movimiento(self, row, col, board):
         possibles = []
-
         # Direcciones diagonales: arriba-derecha, arriba-izquierda, abajo-derecha, abajo-izquierda
         directions = [(1, 1), (1, -1), (-1, 1), (-1, -1)]