+++ piezas/pawn.py
@@ -3,25 +3,32 @@ from piezas.piece import Piece
 
 
 class Pawn(Piece):
+
+    white_str = "♙"
+    black_str = "♟"
     def __init__(self, color):
         super().__init__(color)
 
-    def movimiento_hacia_adelante(self, row, col, es_primer_movimiento, board):
+    #Verifica los movimientos del peon
+    def movimiento(self, row, col, board):
+        
+        es_primer_movimiento = self.__primer_movimiento__
         possibles = []
 
-        # Definir la dirección en función del color
+        #Definir la dirección en función del color
         if self.__color__ == "WHITE":
-            direction = 1  # Los peones blancos avanzan hacia filas mayores
+            direction = -1  #Los peones blancos avanzan hacia filas mayores
         else:  # 'negro'
-            direction = -1  # Los peones negros avanzan hacia filas menores
+            direction = 1  #Los peones negros avanzan hacia filas menores
 
-        # Avance de una casilla
+        #Avance de una casilla
         next_row = row + direction
-        if 0 <= next_row <= 7 and board.get_piece(next_row, col) is None:  # Solo avanza si la casilla está vacía
+        if 0 <= next_row <= 7 and board.get_piece(next_row, col) is None:  #Solo avanza si la casilla está vacía
             possibles.append((next_row, col))
 
-        # Avance de dos casillas si es el primer movimiento
+        #Avance de dos casillas si es el primer movimiento
         if es_primer_movimiento and board.get_piece(next_row, col) is None:
+            
             second_row = row + 2 * direction
             if 0 <= second_row <= 7 and board.get_piece(second_row, col) is None:
                 possibles.append((second_row, col))